+++ internal/pkg/git/git_test.go
@@ -1,23 +1,9 @@
-// nolint: forbidigo
-package git_test
+package git
 
 import (
-	"bytes"
-	"context"
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"testing"
-	"time"
 
-	"github.com/keboola/go-utils/pkg/wildcards"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
-	"github.com/keboola/keboola-as-code/internal/pkg/filesystem/aferofs"
-	. "github.com/keboola/keboola-as-code/internal/pkg/git"
-	"github.com/keboola/keboola-as-code/internal/pkg/log"
 )
 
 func TestGit_Available(t *testing.T) {
@@ -25,142 +11,4 @@ func TestGit_Available(t *testing.T) {
 
 	// should be always true as git is available in the container running the tests
 	assert.True(t, Available())
-}
-
-func TestGit_Checkout(t *testing.T) {
-	t.Parallel()
-	logger := log.NewDebugLogger()
-
-	// Copy the git repository to temp
-	tmpDir := t.TempDir()
-	assert.NoError(t, aferofs.CopyFs2Fs(nil, filepath.Join("test", "repository"), nil, tmpDir))
-	assert.NoError(t, os.Rename(filepath.Join(tmpDir, ".gittest"), filepath.Join(tmpDir, ".git")))
-	gitRepo := fmt.Sprintf("file://%s", tmpDir)
-
-	// Ctx
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	// Checkout fail from a missing repo
-	url := "file://some/missing/repo"
-	ref := "main"
-	_, err := Checkout(ctx, url, ref, true, logger)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), `git repository could not be checked out from "file://some/missing/repo"`)
-
-	// Checkout fail from a non-existing branch
-	url = gitRepo
-	ref = "non-existing-ref"
-	_, err = Checkout(ctx, url, ref, false, logger)
-	assert.Error(t, err)
-	wildcards.Assert(t, `reference "non-existing-ref" not found in the git repository "%s"`, err.Error(), "unexpected output")
-
-	// Success
-	url = gitRepo
-	ref = "main"
-	r, err := Checkout(ctx, url, ref, false, logger)
-	assert.NoError(t, err)
-
-	// Get repository FS
-	fs1, unlockFS1 := r.Fs()
-
-	// Full checkout -> directory is not empty
-	subDirs, err := filesystem.ReadSubDirs(fs1, "/")
-	assert.NoError(t, err)
-	assert.Greater(t, len(subDirs), 1)
-
-	// Check if the hash equals to a commit - the git command should return a "commit" message
-	hash, err := r.CommitHash(ctx)
-	assert.NoError(t, err)
-	var stdOut bytes.Buffer
-	cmd := exec.Command("git", "cat-file", "-t", hash)
-	cmd.Dir = fs1.BasePath()
-	cmd.Stdout = &stdOut
-	err = cmd.Run()
-	assert.NoError(t, err)
-	assert.Equal(t, "commit\n", stdOut.String())
-
-	// Test parallel access to FS
-	fs2, unlockFS2 := r.Fs()
-	assert.True(t, fs2.Exists("example-file.txt"))
-
-	// Test pull
-	assert.NoError(t, r.Pull(ctx))
-	assert.True(t, fs1.Exists("example-file.txt"))
-
-	// Test free
-	basePath := fs1.BasePath()
-	unlockFS1()
-	unlockFS2()
-	<-r.Free()
-	assert.NoDirExists(t, basePath)
-}
-
-func TestGit_Checkout_Sparse(t *testing.T) {
-	t.Parallel()
-	logger := log.NewDebugLogger()
-
-	// Copy the git repository to temp
-	tmpDir := t.TempDir()
-	assert.NoError(t, aferofs.CopyFs2Fs(nil, filepath.Join("test", "repository"), nil, tmpDir))
-	assert.NoError(t, os.Rename(filepath.Join(tmpDir, ".gittest"), filepath.Join(tmpDir, ".git")))
-	gitRepo := fmt.Sprintf("file://%s", tmpDir)
-
-	// Ctx
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	// Checkout fail from a missing repo
-	url := "file://some/missing/repo"
-	ref := "main"
-	_, err := Checkout(ctx, url, ref, true, logger)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), `git repository could not be checked out from "file://some/missing/repo"`)
-
-	// Checkout fail from a non-existing branch
-	url = gitRepo
-	ref = "non-existing-ref"
-	_, err = Checkout(ctx, url, ref, true, logger)
-	assert.Error(t, err)
-	wildcards.Assert(t, `reference "non-existing-ref" not found in the git repository "%s"`, err.Error(), "unexpected output")
-
-	// Success
-	url = gitRepo
-	ref = "main"
-	r, err := Checkout(ctx, url, ref, true, logger)
-	assert.NoError(t, err)
-
-	// Get repository FS
-	fs1, unlockFS1 := r.Fs()
-
-	// Sparse checkout -> directory is empty
-	subDirs, err := filesystem.ReadSubDirs(fs1, "/")
-	assert.NoError(t, err)
-	assert.Equal(t, []string{".git"}, subDirs)
-
-	// Check if the hash equals to a commit - the git command should return a "commit" message
-	hash, err := r.CommitHash(ctx)
-	assert.NoError(t, err)
-	var stdOut bytes.Buffer
-	cmd := exec.Command("git", "cat-file", "-t", hash)
-	cmd.Dir = fs1.BasePath()
-	cmd.Stdout = &stdOut
-	err = cmd.Run()
-	assert.NoError(t, err)
-	assert.Equal(t, "commit\n", stdOut.String())
-
-	// Test parallel access to FS
-	fs2, unlockFS2 := r.Fs()
-	assert.True(t, fs2.Exists(".git"))
-
-	// Test pull
-	assert.NoError(t, r.Pull(ctx))
-	assert.True(t, fs1.Exists(".git"))
-
-	// Test free
-	basePath := fs1.BasePath()
-	unlockFS1()
-	unlockFS2()
-	<-r.Free()
-	assert.NoDirExists(t, basePath)
 }
